# docs/swagger.yaml
basePath: /
definitions:
  domain.ApiKey:
    properties:
      api_key_id:
        description: ID is the primary key of the API key
        type: integer
      api_key_user_id:
        description: USerID references the user who the API key belongs to
        type: integer
      key:
        description: Key is the actual API key string
        type: string
      role:
        description: Role specifies the role associated with the API key
        type: string
    type: object
  domain.Resource:
    properties:
      can_download:
        type: boolean
      current_files:
        items:
          $ref: '#/definitions/domain.ResourceFile'
        type: array
      custom_fields: {}
      description:
        type: string
      download_count:
        type: integer
      external_url:
        type: string
      icon_url:
        type: string
      last_update:
        type: integer
      license:
        type: string
      license_url:
        type: string
      rating_avg:
        type: number
      rating_count:
        type: integer
      rating_weighted:
        type: number
      resource_category_id:
        type: integer
      resource_date:
        type: integer
      resource_id:
        type: integer
      resource_state:
        type: string
      resource_type:
        type: string
      review_count:
        type: integer
      tag_line:
        type: string
      title:
        type: string
      update_count:
        type: integer
      version:
        type: string
      view_count:
        description: CurrentDownloadUrl string      `json:"current_download_url,omitempty"`
        type: integer
      view_url:
        type: string
    type: object
  domain.ResourceCategory:
    properties:
      description:
        type: string
      display_order:
        type: integer
      last_update:
        type: integer
      parent_category_id:
        type: integer
      resource_category_id:
        type: integer
      resource_count:
        type: integer
      title:
        type: string
    type: object
  domain.ResourceFile:
    properties:
      filename:
        type: string
      id:
        type: integer
      size:
        type: integer
    type: object
  domain.ResourceReview:
    properties:
      message:
        type: string
      rating:
        type: integer
      rating_date:
        type: integer
      rating_state:
        type: string
      resource_id:
        type: integer
      resource_rating_id:
        type: integer
      resource_version_id:
        type: integer
    type: object
  domain.ResourceVersion:
    properties:
      download_count:
        type: integer
      files:
        items:
          $ref: '#/definitions/domain.ResourceFile'
        type: array
      release_date:
        type: integer
      resource_id:
        type: integer
      resource_version_id:
        type: integer
      version_string:
        type: string
    type: object
  domain.Server:
    properties:
      description:
        type: string
      has_password:
        type: boolean
      icon_url:
        type: string
      ip:
        type: string
      is_visible:
        type: boolean
      max_clients:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      port:
        type: integer
      server_date:
        type: integer
      server_id:
        type: integer
      server_state:
        $ref: '#/definitions/domain.ServerStatus'
      version:
        type: string
    required:
    - description
    - has_password
    - ip
    - is_visible
    - max_clients
    - name
    - port
    - version
    type: object
  domain.ServerStatus:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - StatusOnline
    - StatusOffline
    - StatusHidden
    - StatusCrashed
  domain.User:
    properties:
      avatar_urls: {}
      email:
        type: string
      gravatar:
        type: string
      is_staff:
        type: boolean
      profile_banner_urls: {}
      user_id:
        type: integer
      user_title:
        type: string
      username:
        type: string
      view_url:
        type: string
    type: object
  remote.RawUserAuthResponse:
    properties:
      login_token:
        type: string
      refresh_token:
        type: string
      tfa_providers:
        type: string
      tfa_required:
        type: boolean
      tfa_triggered:
        type: boolean
      user:
        $ref: '#/definitions/domain.User'
    type: object
  router.RequestError:
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    name: Rafael Galvan
    url: http://www.rigsofrods.org
  license:
    name: GNU GPL v3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  title: Rigs of Rods API
  version: "2.0"
paths:
  /api_keys/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ApiKey'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - api_key
    post:
      consumes:
      - application/json
      parameters:
      - description: API Key Object
        in: body
        name: apiKey
        required: true
        schema:
          $ref: '#/definitions/domain.ApiKey'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ApiKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - api_key
  /api_keys/{key}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: API Key Identifier
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - api_key
  /auth/login/:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/remote.RawUserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - auth
  /auth/logout/:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/remote.RawUserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - auth
  /auth/refresh/:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/remote.RawUserAuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - auth
  /resource-categories/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ResourceCategory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resource-categories/{category}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ResourceCategory'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resource-versions/{version}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ResourceVersion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resources/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Resource'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resources/{resource}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Resource'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resources/{resource}/reviews:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ResourceReview'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /resources/{resource}/versions:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ResourceVersion'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - resource
  /servers:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Server'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - server
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - server
    put:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - server
  /servers/{server}:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Server'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - server
  /users/{user}/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - users
  /users/me/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.RequestError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/router.RequestError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.RequestError'
      tags:
      - users
swagger: "2.0"