# router/router_user.go
// Copyright (C) 2022 Rafael Galvan

// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU General Public License as published by
// the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.

// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU General Public License for more details.

// You should have received a copy of the GNU General Public License
// along with this program.  If not, see <http://www.gnu.org/licenses/>.

package router

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// ShowAccount godoc
//	@Tags		users
//	@Accept		json
//	@Produce	json
//	@Success	200	{object}	[]domain.User
//	@Failure	400	{object}	RequestError
//	@Failure	404	{object}	RequestError
//	@Failure	500	{object}	RequestError
//	@Router		/users/{user}/ [get]
func getUser(c *gin.Context) {
	c.Status(http.StatusNotImplemented)
}

// ShowAccount godoc
//	@Tags		users
//	@Accept		json
//	@Produce	json
//	@Success	200	{object}	domain.User
//	@Failure	400	{object}	RequestError
//	@Failure	404	{object}	RequestError
//	@Failure	500	{object}	RequestError
//	@Router		/users/me/ [get]
func getMe(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"me": ExtractUser(c),
	})
}